+++ operations/producer/MessageSerializer.go
@@ -1,0 +1,9 @@
+package producer
+
+import (
+	"github.com/Shopify/sarama"
+)
+
+type MessageSerializer interface {
+	Serialize(key, value []byte, flags ProducerFlags) *sarama.ProducerMessage
+}